1. splashscreen 수정 및 dart생성
