+++ lib/main.dart
@@ -1,65 +1,28 @@
 import 'package:flutter/material.dart';
 import 'properties/InformationProperties.dart';
-import 'dart:async';
-import 'view/root_screen.dart';
+import 'view/screen/splash_screen.dart';
 
-void main() {
-  runApp(new MaterialApp(
-    title: InformationProperties.APP_NAME,
-    theme: ThemeData(
-      // This is the theme of your application.
-      //
-      // Try running your application with "flutter run". You'll see the
-      // application has a blue toolbar. Then, without quitting the app, try
-      // changing the primarySwatch below to Colors.green and then invoke
-      // "hot reload" (press "r" in the console where you ran "flutter run",
-      // or simply save your changes to "hot reload" in a Flutter IDE).
-      // Notice that the counter didn't reset back to zero; the application
-      // is not restarted.
-      primarySwatch: Colors.blue,
-    ),
-    home: new SplashScreen(),
-    routes: <String, WidgetBuilder>{
-      '/RootScreen': (BuildContext context) => new RootScreen()
-    },
-  ));
-}
+void main() => runApp(MyApp());
 
-class SplashScreen extends StatefulWidget {
-  @override
-  _SplashScreenState createState() => new _SplashScreenState();
-}
-
-class _SplashScreenState extends State<SplashScreen> {
-  startTime() async {
-    var _duration = new Duration(seconds: 2);
-    return new Timer(_duration, navigationPage);
-  }
-
-  void navigationPage() {
-    Navigator.of(context).pushReplacementNamed('/RootScreen');
-  }
-
-  @override
-  void initState() {
-    super.initState();
-    startTime();
-  }
-
+class MyApp extends StatelessWidget {
+  // This widget is the root of your application.
   @override
   Widget build(BuildContext context) {
-    return Container(
-      decoration: BoxDecoration(
-        image: DecorationImage(
-            image: AssetImage('assets/splashlogo.png'),
-            fit: BoxFit.cover
-        ) ,
+    return MaterialApp(
+      title: InformationProperties.APP_NAME,
+      theme: ThemeData(
+        // This is the theme of your application.
+        //
+        // Try running your application with "flutter run". You'll see the
+        // application has a blue toolbar. Then, without quitting the app, try
+        // changing the primarySwatch below to Colors.green and then invoke
+        // "hot reload" (press "r" in the console where you ran "flutter run",
+        // or simply save your changes to "hot reload" in a Flutter IDE).
+        // Notice that the counter didn't reset back to zero; the application
+        // is not restarted.
+        primarySwatch: Colors.blue,
       ),
-      child: Center(
-        child: CircularProgressIndicator(
-          valueColor: AlwaysStoppedAnimation<Color>(Colors.redAccent),
-        ),
-      ),
+      home: SplashScreen(),
     );
   }
 }
