+++ lib/view/screen/home_screen.dart
@@ -242,5 +242,4 @@ class HomeScreenState extends State<HomeScreen>{
       ),
     );
   }
-
 }
