+++ lib/view/screen/splash_screen.dart
@@ -1,0 +1,47 @@
+import 'package:flutter/material.dart';
+import 'dart:async';
+import 'package:dambi/view/root_screen.dart';
+
+
+class SplashScreen extends StatefulWidget {
+
+  @override
+  State<StatefulWidget> createState() {
+    return SplashScreenState();
+  }
+}
+
+class SplashScreenState extends State<SplashScreen> {
+
+  @override
+  void initState() {
+    super.initState();
+    loadData();
+  }
+
+  Future<Timer> loadData() async {
+    return new Timer(Duration(seconds: 5), onDoneLoading);
+  }
+
+  onDoneLoading() async {
+    Navigator.of(context).push(MaterialPageRoute(builder: (context) => RootScreen()));
+  }
+
+
+  @override
+  Widget build(BuildContext context) {
+    return Container(
+      decoration: BoxDecoration(
+        image: DecorationImage(
+            image: AssetImage('assets/splashlogo.png'),
+            fit: BoxFit.cover
+        ) ,
+      ),
+      child: Center(
+        child: CircularProgressIndicator(
+          valueColor: AlwaysStoppedAnimation<Color>(Colors.redAccent),
+        ),
+      ),
+    );
+  }
+}
